+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/Cron/Cron.styles.ts
@@ -1,6 +1,12 @@
-
 import { Button } from '@mui/material';
 import styled from 'styled-components';
+
+export const CronContainer = styled.div`
+    font-family: 'roboto';
+    display: flex;
+    align-items: flex-start;
+    flex-wrap: wrap;
+`;
 
 export const ClearButton = styled(Button)(({ theme }) => `
     && {
@@ -11,10 +17,4 @@ export const ClearButton = styled(Button)(({ theme }) => `
             background-color: ${theme.palette.common.black};
         }
     }
-`,
-);
-
-export const PeriodText = styled.span`
-    padding-left: 5px;
-    font-weight: bold;
-`;
+`);
