+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/Cron/index.ts
@@ -1,0 +1,3 @@
+import Cron from './Cron';
+
+export default Cron;
