+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/components/CustomSelect/CustomSelect.styles.ts
@@ -1,0 +1,9 @@
+import { Select } from '@mui/material';
+import styled from 'styled-components';
+
+export const StyledSelect = styled(Select)(({ theme }) => `
+    && { 
+        margin-left: ${theme.spacing(1)};
+        padding-inline: ${theme.spacing(2)} 0;
+    }
+`);
