+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/components/CustomSelect/index.ts
@@ -1,0 +1,3 @@
+import CustomSelect from './CustomSelect';
+
+export default CustomSelect;
