+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/components/PeriodDefinition/PeriodDefinition.styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const PeriodText = styled.span(({ theme }) => `
+    margin-left: ${theme.spacing(1)};
+    font-weight: ${theme.typography.fontWeightBold};
+`);
