+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/components/PeriodDefinition/PeriodDefinition.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
-import { PeriodText } from '../../Cron.styles';
 import DEFAULT_LOCALE_EN from '../../locale';
+import { PeriodText } from './PeriodDefinition.styles';
 import { PeriodDefinitionProps } from './PeriodDefinition.types';
 
 const PeriodDefinition: FC<PeriodDefinitionProps>= ({
