+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/MonthDaysSection/NthMonthDaysSelect/NthMonthDaysSelect.tsx
@@ -16,7 +16,7 @@ const NthMonthDaysSelect: FC<NthMonthDaysSelectProps>= ({
     period,
 }) => {
     const optionsList = locale.nthMonthDays || DEFAULT_LOCALE_EN.nthMonthDays;
-    
+
     return (
         <CustomSelect
             period={period}
