+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Period/Period.types.ts
@@ -1,0 +1,7 @@
+import { FieldProps, PeriodType, SetValuePeriod, Shortcuts } from '../../types';
+
+export interface PeriodProps extends Omit<FieldProps, 'value' | 'setValue' | 'period'> {
+    value: PeriodType;
+    setValue: SetValuePeriod;
+    shortcuts: Shortcuts;
+}
