+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Period/index.ts
@@ -1,0 +1,3 @@
+import Period from './Period';
+
+export default Period;
