+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/locale.ts
@@ -1,6 +1,6 @@
-import { translate } from '#src-app/hooks/useTranslations';
-
 import { DefaultLocale } from './types';
+
+import { translate } from '#src-app/hooks/useTranslations';
 
 const DEFAULT_LOCALE_EN: DefaultLocale = {
     everyText: translate('Component.Cron.Locale.Every'),
